% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_coding.R
\name{phase_ratio}
\alias{phase_ratio}
\title{phase_ratio}
\usage{
phase_ratio(pa)
}
\arguments{
\item{pa}{phase-angle vector.}
}
\value{
A phase-ratio dataframe.
}
\description{
phase-ratio statistics
}
\examples{
data(sample)
v1 <- sample$hip
v2 <- sample$knee
v3 <- sample$ankle
pa <- phase_angle(v1, v2)
phase_ratio(pa)
}
